correct bad behavior on descriptions generation
