+++ lib/api/controllers/remoteActions/swagger.js
@@ -8,6 +8,7 @@ var
   generateSwagger = require('../../core/swagger'),
   writeFiles;
 
+
 writeFiles = swagger => {
   var baseDir = path.join(__dirname, '..', '..', '..', '..');
 
@@ -35,20 +36,6 @@ module.exports = function GenerateSwaggerFiles (kuzzle) {
     return q({isWorker: true});
   }
 
-  kuzzle.config.httpRoutes.forEach(_route => {
-    var route = _.assign({}, _route);
-    route.url = '/' + kuzzle.config.apiVersion + route.url;
-    routes.push(route);
-  });
-
-  routes.push({verb: 'get', url: '/_serverInfo', controller: 'read', action: 'serverInfo'});
-
-  kuzzle.pluginsManager.routes.forEach(_route => {
-    var route = _.assign({}, _route);
-    route.url = '/' + kuzzle.config.apiVersion + '/_plugin' + route.url;
-    routes.push(route);
-  });
-
-  return generateSwagger(routes, kuzzle.config.apiVersion)
+  return generateSwagger(kuzzle)
     .then(swagger => writeFiles(swagger));
 };
