+++ lib/api/controllers/routerController.js
@@ -227,7 +227,7 @@ module.exports = function RouterController (kuzzle) {
     });
 
     apiBase.get('/swagger.json', (request, response) => {
-      generateSwagger(routes, kuzzle.config.apiVersion)
+      generateSwagger(kuzzle)
         .then(swagger => {
           response.setHeader('Access-Control-Allow-Origin', '*');
           response.setHeader('Access-Control-Allow-Headers', 'X-Requested-With');
@@ -237,7 +237,7 @@ module.exports = function RouterController (kuzzle) {
     });
 
     apiBase.get('/swagger.yml', (request, response) => {
-      generateSwagger(routes, kuzzle.config.apiVersion)
+      generateSwagger(kuzzle)
         .then(swagger => {
           response.setHeader('Access-Control-Allow-Origin', '*');
           response.setHeader('Access-Control-Allow-Headers', 'X-Requested-With');
