+++ lib/api/core/swagger.js
@@ -1,5 +1,6 @@
 var
-  q = require('q');
+  q = require('q'),
+  _ = require('lodash');
 
 /**
  * Generates JSON swagger object
@@ -9,8 +10,23 @@ var
  *
  * @returns {Promise} promise
  */
-module.exports = function generateSwagger (routes, apiVersion) {
-  var swagger;
+module.exports = function generateSwagger (kuzzle) {
+  var swagger, apiVersion, routes = [];
+  apiVersion = kuzzle.config.apiVersion;
+
+  kuzzle.config.httpRoutes.forEach(_route => {
+    var route = _.assign({}, _route);
+    route.url = '/' + apiVersion + route.url;
+    routes.push(route);
+  });
+
+  routes.push({verb: 'get', url: '/_serverInfo', controller: 'read', action: 'serverInfo'});
+
+  kuzzle.pluginsManager.routes.forEach(_route => {
+    var route = _.assign({}, _route);
+    route.url = '/' + apiVersion + '/_plugin' + route.url;
+    routes.push(route);
+  });  
 
   swagger = {
     basePath: '/api',
